# Conflicts:
#	build.gradle
#	demo/src/main/java/com/google/android/exoplayer/demo/EventLogger.java
#	demo/src/main/java/com/google/android/exoplayer/demo/PlayerActivity.java
#	demo/src/main/java/com/google/android/exoplayer/demo/player/DashRendererBuilder.java
#	demo/src/main/java/com/google/android/exoplayer/demo/player/DemoPlayer.java
#	library/src/main/java/com/google/android/exoplayer/ExoPlayerImpl.java
#	library/src/main/java/com/google/android/exoplayer/ExoPlayerImplInternal.java
#	library/src/main/java/com/google/android/exoplayer/MediaCodecUtil.java
#	library/src/main/java/com/google/android/exoplayer/audio/AudioTrack.java
#	library/src/main/java/com/google/android/exoplayer/dash/DashChunkSource.java
#	library/src/main/java/com/google/android/exoplayer/dash/mpd/MediaPresentationDescriptionParser.java
#	library/src/main/java/com/google/android/exoplayer/extractor/mp4/FragmentedMp4Extractor.java
#	library/src/main/java/com/google/android/exoplayer/extractor/mp4/PsshAtomUtil.java
#	library/src/main/java/com/google/android/exoplayer/extractor/ogg/VorbisUtil.java
#	library/src/main/java/com/google/android/exoplayer/extractor/ts/AdtsReader.java
#	library/src/main/java/com/google/android/exoplayer/extractor/ts/H265Reader.java
#	library/src/main/java/com/google/android/exoplayer/extractor/ts/TsExtractor.java
#	library/src/main/java/com/google/android/exoplayer/extractor/wav/WavHeaderReader.java
#	library/src/main/java/com/google/android/exoplayer/hls/HlsChunkSource.java
#	library/src/main/java/com/google/android/exoplayer/text/CuePainter.java
#	library/src/main/java/com/google/android/exoplayer/text/subrip/SubripParser.java
#	library/src/main/java/com/google/android/exoplayer/text/ttml/TtmlParser.java
#	library/src/main/java/com/google/android/exoplayer/text/webvtt/WebvttCue.java
#	library/src/main/java/com/google/android/exoplayer/text/webvtt/WebvttCueParser.java
#	library/src/main/java/com/google/android/exoplayer/upstream/DefaultHttpDataSource.java
#	library/src/main/java/com/google/android/exoplayer/upstream/Loader.java
#	library/src/main/java/com/google/android/exoplayer/upstream/UdpDataSource.java
#	library/src/main/java/com/google/android/exoplayer/upstream/cache/CacheDataSource.java
#	library/src/main/java/com/google/android/exoplayer/util/NalUnitUtil.java

+++ library/src/main/java/com/google/android/exoplayer/hls/HlsMediaPlaylist.java
@@ -16,7 +16,6 @@
 package com.google.android.exoplayer.hls;
 
 import com.google.android.exoplayer.C;
-
 import java.util.List;
 
 /**
@@ -36,12 +35,12 @@ public final class HlsMediaPlaylist extends HlsPlaylist {
     public final boolean isEncrypted;
     public final String encryptionKeyUri;
     public final String encryptionIV;
-    public final int byterangeOffset;
-    public final int byterangeLength;
+    public final long byterangeOffset;
+    public final long byterangeLength;
 
     public Segment(String uri, double durationSecs, int discontinuitySequenceNumber,
         long startTimeUs, boolean isEncrypted, String encryptionKeyUri, String encryptionIV,
-        int byterangeOffset, int byterangeLength) {
+        long byterangeOffset, long byterangeLength) {
       this.url = uri;
       this.durationSecs = durationSecs;
       this.discontinuitySequenceNumber = discontinuitySequenceNumber;
